requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
fastapi==0.70.1
h11==0.12.0
httptools==0.3.0
idna==3.3
itsdangerous==2.0.1
oauthlib==3.1.1
pydantic==1.8.2
python-dotenv==0.19.2
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
sniffio==1.2.0
starlette==0.16.0
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
